# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Healthy Nation API — vitals, doctors, medications, AI triage, emergency, pharmacy
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /doctors:
    get:
      operationId: listDoctors
      tags: [doctors]
      summary: Search doctors
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: specialty
          schema:
            type: string
        - in: query
          name: minRating
          schema:
            type: number
        - in: query
          name: visitType
          schema:
            $ref: "#/components/schemas/VisitType"
      responses:
        "200":
          description: Doctors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Doctor"
  /doctors/{id}:
    get:
      operationId: getDoctor
      tags: [doctors]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Doctor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"
  /doctors/{id}/availability:
    get:
      operationId: getDoctorAvailability
      tags: [doctors]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: query
          name: days
          schema:
            type: integer
            default: 7
      responses:
        "200":
          description: Availability
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DayAvailability"
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      parameters:
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/AppointmentStatus"
      responses:
        "200":
          description: Appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    patch:
      operationId: updateAppointment
      tags: [appointments]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /medications:
    get:
      operationId: listMedications
      tags: [medications]
      parameters:
        - in: query
          name: ongoing
          schema:
            type: boolean
      responses:
        "200":
          description: Medications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Medication"
    post:
      operationId: createMedication
      tags: [medications]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMedicationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medication"
  /medications/{id}:
    patch:
      operationId: updateMedication
      tags: [medications]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMedicationInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medication"
    delete:
      operationId: deleteMedication
      tags: [medications]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /conditions:
    get:
      operationId: listConditions
      tags: [conditions]
      responses:
        "200":
          description: Conditions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Condition"
    post:
      operationId: createCondition
      tags: [conditions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConditionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Condition"
  /conditions/{id}:
    delete:
      operationId: deleteCondition
      tags: [conditions]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /emergency/contacts:
    get:
      operationId: listEmergencyContacts
      tags: [emergency]
      responses:
        "200":
          description: Contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmergencyContact"
    post:
      operationId: createEmergencyContact
      tags: [emergency]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmergencyContactInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergencyContact"
  /emergency/contacts/{id}:
    delete:
      operationId: deleteEmergencyContact
      tags: [emergency]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /emergency/sos:
    post:
      operationId: triggerSos
      tags: [emergency]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SosRequest"
      responses:
        "201":
          description: SOS event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SosEvent"
...
  /emergency/sos/{id}/resolve:
    post:
      operationId: resolveSos
      tags: [emergency]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SosEvent"
components:
  schemas:
...
    VisitType:
      type: string
      enum: [telemedicine, in_person, phone]
    AppointmentStatus:
      type: string
      enum: [pending_confirmation, confirmed, in_progress, completed, cancelled, no_show]
...
    Doctor:
      type: object
      required: [id, name, specialty, bio, rating, ratingCount, experienceYears, languages, feeInPerson, feeTelemedicine, supportsTelemedicine, clinicName, city, photoUrl, verified]
      properties:
        id:
          type: integer
        name:
          type: string
        specialty:
          type: string
        bio:
          type: string
        rating:
          type: number
        ratingCount:
          type: integer
        experienceYears:
          type: integer
        languages:
          type: array
          items:
            type: string
        feeInPerson:
          type: integer
        feeTelemedicine:
          type: integer
        supportsTelemedicine:
          type: boolean
        clinicName:
          type: string
        city:
          type: string
        photoUrl:
          type: string
        verified:
          type: boolean
        nextAvailable:
          type: ["string", "null"]
...
    DayAvailability:
      type: object
      required: [date, slots]
      properties:
        date:
          type: string
        slots:
          type: array
          items:
            $ref: "#/components/schemas/Slot"
    Slot:
      type: object
      required: [time, available]
      properties:
        time:
          type: string
        available:
          type: boolean
    Appointment:
      type: object
      required: [id, doctorId, doctorName, doctorSpecialty, doctorPhotoUrl, scheduledAt, visitType, status, fee, reason]
      properties:
        id:
          type: integer
        doctorId:
          type: integer
        doctorName:
          type: string
        doctorSpecialty:
          type: string
        doctorPhotoUrl:
          type: string
        scheduledAt:
          type: string
          format: date-time
        visitType:
          $ref: "#/components/schemas/VisitType"
        status:
          $ref: "#/components/schemas/AppointmentStatus"
        fee:
          type: integer
        reason:
          type: string
        symptoms:
          type: ["string", "null"]
        videoCallLink:
          type: ["string", "null"]
    CreateAppointmentInput:
      type: object
      required: [doctorId, scheduledAt, visitType, reason]
      properties:
        doctorId:
          type: integer
        scheduledAt:
          type: string
          format: date-time
        visitType:
          $ref: "#/components/schemas/VisitType"
        reason:
          type: string
        symptoms:
          type: ["string", "null"]
    UpdateAppointmentInput:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/AppointmentStatus"
        cancellationReason:
          type: ["string", "null"]
    Medication:
      type: object
      required: [id, name, dosage, frequency, indication, startDate, isOngoing, refillRemaining, adherencePercent]
      properties:
        id:
          type: integer
        name:
          type: string
        dosage:
          type: string
        frequency:
          type: string
        indication:
          type: string
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        isOngoing:
          type: boolean
        refillRemaining:
          type: integer
        adherencePercent:
          type: integer
        timeOfDay:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    CreateMedicationInput:
      type: object
      required: [name, dosage, frequency, indication, startDate]
      properties:
        name:
          type: string
        dosage:
          type: string
        frequency:
          type: string
        indication:
          type: string
        startDate:
          type: string
        timeOfDay:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    UpdateMedicationInput:
      type: object
      properties:
        dosage:
          type: ["string", "null"]
        frequency:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isOngoing:
          type: ["boolean", "null"]
    Condition:
      type: object
      required: [id, name, status, severity, diagnosisDate]
      properties:
        id:
          type: integer
        name:
          type: string
        status:
          type: string
          enum: [active, inactive, resolved]
        severity:
          type: string
          enum: [mild, moderate, severe]
        diagnosisDate:
          type: string
        notes:
          type: ["string", "null"]
        icdCode:
          type: ["string", "null"]
    CreateConditionInput:
      type: object
      required: [name, status, severity, diagnosisDate]
      properties:
        name:
          type: string
        status:
          type: string
          enum: [active, inactive, resolved]
        severity:
          type: string
          enum: [mild, moderate, severe]
        diagnosisDate:
          type: string
        notes:
          type: ["string", "null"]
        icdCode:
          type: ["string", "null"]
...
    EmergencyContact:
      type: object
      required: [id, name, relationship, phone, isPrimary]
      properties:
        id:
          type: integer
        name:
          type: string
        relationship:
          type: string
        phone:
          type: string
        isPrimary:
          type: boolean
    CreateEmergencyContactInput:
      type: object
      required: [name, relationship, phone]
      properties:
        name:
          type: string
        relationship:
          type: string
        phone:
          type: string
        isPrimary:
          type: ["boolean", "null"]
    Hospital:
      type: object
      required: [id, name, address, distanceKm, phone, traumaLevel, erOpen, etaMinutes]
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
        distanceKm:
          type: number
        phone:
          type: string
        traumaLevel:
          type: string
        erOpen:
          type: boolean
        etaMinutes:
          type: integer
    SosRequest:
      type: object
      properties:
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
    SosEvent:
      type: object
      required: [id, status, triggeredAt, contactsNotified, hospitals]
      properties:
        id:
          type: integer
        status:
          type: string
          enum: [active, resolved, cancelled]
        triggeredAt:
          type: string
          format: date-time
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        contactsNotified:
          type: integer
        hospitals:
          type: array
          items:
            $ref: "#/components/schemas/Hospital"