# lib/api-zod/src/generated/types/sosRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Healthy Nation API — vitals, doctors, medications, AI triage, emergency, pharmacy
 * OpenAPI spec version: 0.1.0
 */

export interface SosRequest {
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  /** @nullable */
  notes?: string | null;
}